+++ components/ui/index.ts
@@ -1,5 +1,4 @@
 import { CursiveTitle } from './CursiveTitle';
-import { Input } from '@/components/ui/Input';
 import { ButtonLink } from '@/components/ui/ButtonLink';
 import { Title2, Title3, Title4, Title5 } from '@/components/ui/Titles'
 
@@ -11,5 +10,4 @@ export {
   Title5,
   CursiveTitle,
   ButtonLink,
-  Input,
 }
