+++ next-auth.d.ts
@@ -1,7 +1,7 @@
-import NextAuth, { type DefaultSession } from "next-auth";
+import NextAuth, { type DefaultSession } from "next-auth";;
 
 export type ExtendedUser = DefaultSession["user"] & {
-  role: 'user' | 'admin'
+  role: 'user' | 'admin';
 };
 
 declare module "next-auth" {
